Modify Task.API endpoints

+++ src/Task.API/Endpoints/TaskItemEndpoints.cs
@@ -18,21 +18,33 @@ public class TaskItemEndpoints : ICarterModule
         group.MapDelete("/{id:guid}", DeleteTask);
     }
 
-    public static async Task<IResult> GetAllTasks(TaskContext context, IMapper mapper)
+    public static async Task<Results<Ok<IEnumerable<TaskItemDto>>, BadRequest<string>>> GetAllTasks(
+        TaskContext context,
+        IMapper mapper)
     {
         var tasks = await context.TaskItems.ToListAsync();
 
         return TypedResults.Ok(mapper.Map<IEnumerable<TaskItemDto>>(tasks));
     }
 
-    public static async Task<Results<Ok<TaskItemDto>, BadRequest<string>>> GetTaskById(
+    public static async Task<Results<Ok<TaskItemDto>, NotFound<string>, BadRequest<string>>> GetTaskById(
         TaskContext context,
         IMapper mapper,
         Guid id)
     {
         try
         {
-            var task = await context.TaskItems.FindAsync(id);
+            if (!Guid.TryParse(id.ToString(), out _))
+            {
+                return TypedResults.BadRequest($"'{id}' is not a valid ID!");
+            }
+
+            var task = await context.TaskItems.SingleOrDefaultAsync(t => t.Id == id);
+
+            if (task == null)
+            {
+                return TypedResults.NotFound($"Task with ID {id} not found!");
+            }
 
             return TypedResults.Ok(mapper.Map<TaskItemDto>(task));
         }
@@ -42,19 +54,19 @@ public class TaskItemEndpoints : ICarterModule
         }
     }
 
-    public static async Task<Results<Ok<string>, BadRequest<string>>> AddTask(
+    public static async Task<Results<Created, BadRequest<string>>> AddTask(
         TaskContext context,
-        [FromBody] IMapper mapper,
+        [AsParameters] IMapper mapper,
         TaskItemDto taskDto)
     {
         try
         {
             var task = mapper.Map<TaskItem>(taskDto);
 
-            await context.TaskItems.AddAsync(task);
+            context.TaskItems.Add(task);
             await context.SaveChangesAsync();
 
-            return TypedResults.Ok("Task added successfully!");
+            return TypedResults.Created($"/api/v1/tasks/{taskDto.Id}");
         }
         catch (Exception ex)
         {
@@ -62,19 +74,26 @@ public class TaskItemEndpoints : ICarterModule
         }
     }
 
-    public static async Task<Results<Ok<string>, BadRequest<string>>> UpdateTask(
+    public static async Task<Results<Created, NotFound<string>, BadRequest<string>>> UpdateTask(
         TaskContext context,
-        [FromBody] IMapper mapper,
+        IMapper mapper,
         TaskItemDto taskDto)
     {
         try
         {
-            var task = mapper.Map<TaskItem>(taskDto);
+            var task = await context.TaskItems.SingleOrDefaultAsync(t => t.Id == taskDto.Id);
 
-            context.TaskItems.Update(task);
+            if (task == null)
+            {
+                return TypedResults.NotFound($"Task with ID {taskDto.Id} not found!");
+            }
+
+            var taskEntry = context.Entry(task);
+            taskEntry.CurrentValues.SetValues(taskDto);
+
             await context.SaveChangesAsync();
 
-            return TypedResults.Ok("Task updated successfully!");
+            return TypedResults.Created($"/api/v1/tasks/{taskDto.Id}");
         }
         catch (Exception ex)
         {
@@ -82,19 +101,24 @@ public class TaskItemEndpoints : ICarterModule
         }
     }
 
-    public static async Task<Results<Ok<string>, BadRequest<string>>> DeleteTask(
+    public static async Task<Results<NoContent, NotFound, BadRequest<string>>> DeleteTask(
         TaskContext context,
         IMapper mapper,
         Guid id)
     {
         try
         {
-            var task = await context.TaskItems.FindAsync(id);
+            var task = await context.TaskItems.SingleOrDefaultAsync(t => t.Id == id);
+
+            if (task == null)
+            {
+                return TypedResults.NotFound();
+            }
 
             context.TaskItems.Remove(task);
             await context.SaveChangesAsync();
 
-            return TypedResults.Ok("Task deleted successfully!");
+            return TypedResults.NoContent();
         }
         catch (Exception ex)
         {
